requirements.txt
openai>=1.0.0
pdfminer.six>=20221105
dotenv>=0.0.5
pydantic>=1.10.0
pymupdf>=1.18.0